+++ en.js
@@ -556,7 +556,7 @@ export default {
   'plans.title': 'Get more out of Literary Universe!',
   'plans.text': `Literary Universe offers  basic functionality for free. If you are avid reader or a creator you might find our advance membership more to your liking.`,
   'plans.faq.basicsPriced': 'Why has the free membership so few things?',
-  'plans.faq.justification': `Here at Literary Universe we value your privacy and intellectual property. As such we don't use your data to sell adds or provide them to other companies. We don't provide any adds of our own either (we don't want any distractions). This means that we have to build many things on our own. This takes a lot of time and is expensive to run. So we have established the free membership as a small preview of Literary Universe with multiple membership tiers to accommodate everyone's needs.`, // TODO expand
+  'plans.faq.justification': `Here at Literary Universe we value your privacy and intellectual property. As such we don't use your data to sell adds or provide them to other companies. We don't provide any adds of our own either (we don't want any distractions). This means that we have to build many things on our own. This takes a lot of time and is expensive to run. So we have established the free membership as a small preview of Literary Universe (that still allows you to do the most important things) with multiple membership tiers and addons to accommodate everyone's needs while at the same time allowing us to keep the service up and continue development.`,
   'plans.limits': 'Limits',
   'plans.features': 'Features',
   'plans.account.visitor': 'Visitor',
@@ -577,13 +577,10 @@ export default {
   'plans.price.perMonth': ' / month',
   'plans.price.perYear': ' / year',
   'plans.development': 'In development',
-  'plans.feature.scratchpad.pads': `{num, plural,
-    one {1 pad}
-    other {# pads}
-  }`,
-  'plans.limits.universes': 'Universes',
+  'plans.limits.encEntries': 'Encyclopedia entries',
   'plans.limits.scratchpadPads': 'Scratchpad pads',
   'plans.limits.maxCollaborators': 'Collaborators per project',
+  'plans.limits.pmChats': 'Maximum # of recipients for private message',
   'plans.feature.encCustomFields': 'Customizable encyclopedia fields on an entry',
   'plans.feature.encAdvCategories': 'More encyclopedia categories',
   'plans.feature.scratchpad': 'Scratchpad',
@@ -594,6 +591,7 @@ export default {
   'plans.feature.challenges': 'Writing challenges',
   'plans.feature.newFeaturesVote': 'Vote on new features',
   'plans.feature.newFeaturesSuggest': 'Propose new features',
+  'plans.feature.accessToBeta': 'Access Beta features',
 
   'nanowrimo.full': 'National Novel Writing Month',
   'nanowrimo.site': 'NaNoWriMo official site',
@@ -1080,7 +1078,9 @@ export default {
   'settings.links.title': 'Link to your other websites',
   'settings.links.description': 'These links will be displayed on your profile page.',
   'settings.links.website': 'Your personal website',
-  'settings.links.social': 'Other websites'
+  'settings.links.social': 'Other websites',
+  // Added 2019-11-07
+  'profile.website': 'My personal website'
 };
 
 /* eslint-enable max-len */